Refactor tests to use confidentialAsset directly and streamline imports across the codebase
