+++ confidential-assets/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  ...require('../jest.config.js'),
+  globalSetup: undefined,
+  globalTeardown: undefined,
+};
