+++ confidential-assets/package.json
@@ -12,5 +12,11 @@
     "@noble/curves": "^1.6.0",
     "@noble/hashes": "^1.5.0",
     "@aptos-labs/ts-sdk": "^1.38.0"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.14",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.8.2"
   }
 }
