+++ confidential-assets/tests/units/checkBalances.test.ts
@@ -1,6 +1,5 @@
-import { getBalances, getTestAccount, getTestConfidentialAccount } from "../helpers";
-import { preloadTables } from "../kangaroo/wasmPollardKangaroo";
-import { longTestTimeout } from "../../helper";
+import { getBalances, getTestAccount, getTestConfidentialAccount, longTestTimeout } from "../helpers";
+import { preloadTables } from "../helpers/wasmPollardKangaroo";
 
 describe("Check balance", () => {
   const alice = getTestAccount();
