+++ confidential-assets/tests/units/checkIsNormalized.test.ts
@@ -1,10 +1,10 @@
-import { aptos, getTestAccount, TOKEN_ADDRESS } from "../helpers";
+import { confidentialAsset, getTestAccount, TOKEN_ADDRESS } from "../helpers";
 
 describe("Check is normalized", () => {
   const alice = getTestAccount();
 
   it("should check if user confidential balance is normalized", async () => {
-    const isAliceBalanceNormalized = await aptos.confidentialAsset.isUserBalanceNormalized({
+    const isAliceBalanceNormalized = await confidentialAsset.isUserBalanceNormalized({
       accountAddress: alice.accountAddress,
       tokenAddress: TOKEN_ADDRESS,
     });
