+++ confidential-assets/tests/units/checkIsRegistered.test.ts
@@ -1,10 +1,10 @@
-import { aptos, getTestAccount, TOKEN_ADDRESS } from "../helpers";
+import { confidentialAsset, getTestAccount, TOKEN_ADDRESS } from "../helpers";
 
 describe("Check Registration status", () => {
   const alice = getTestAccount();
 
   it("should return true if the user is registered", async () => {
-    const isAliceRegistered = await aptos.confidentialAsset.hasUserRegistered({
+    const isAliceRegistered = await confidentialAsset.hasUserRegistered({
       accountAddress: alice.accountAddress,
       tokenAddress: TOKEN_ADDRESS,
     });
