+++ confidential-assets/tests/units/getAssetAuditor.test.ts
@@ -1,8 +1,8 @@
-import { aptos, TOKEN_ADDRESS } from "../helpers";
+import { confidentialAsset, TOKEN_ADDRESS } from "../helpers";
 
 describe("Global auditor", () => {
   it("it should get global auditor", async () => {
-    const [{ vec }] = await aptos.confidentialAsset.getAssetAuditor({
+    const [{ vec }] = await confidentialAsset.getAssetAuditor({
       tokenAddress: TOKEN_ADDRESS,
     });
